switch to "po4a" command
The upstream has deprecated po4a-translate and updatepo, in favor of the
"po4a" command.

We generate po4a.cfg as-needed with a Python script.

The advantage of "po4a" is the msgstr of same msgid in different files
are synced, while the disadvantage is it's slower (like using "-flto
-Wodr" to find ODR violations with GCC).